pyproject.toml
[tool.poetry]
name = "api-project-00"
version = "0.1.0"
description = ""
authors = ["enterspacebs <anotherhell48@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "api_project_00"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
openai = "^0.27.8"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
